+++ CHANGELOG.md
@@ -103,6 +103,7 @@
 - Scripts: added AppleScripts to make BittyTax easier to use on macOS.
 - Coinbase parser: added "Subscription Rebates (24 Hours)" transaction type.
 - Conversion tool: added new parser for Neon Wallet.
+- Config: added optional API key config for data sources.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))